+++ packages/tx-tool/src/notaryKeyPair.ts
@@ -1,10 +1,10 @@
-import { Signature } from '@radixdlt/radix-engine-toolkit';
+import { Signature } from '@steleaio/radix-engine-toolkit';
 import { Array as A, Effect, flow, Option } from 'effect';
 import type { HexString } from '@radix-effects/shared';
 import { Signer } from './signer/signer';
 
 export class NotaryKeyPair extends Effect.Service<NotaryKeyPair>()(
-  'NotaryKeyPair',
+  '@radix-effects/tx-tool/NotaryKeyPair',
   {
     effect: Effect.gen(function* () {
       const signer = yield* Signer;