+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @radix-effects/gateway
 
+## 0.6.0
+
+### Minor Changes
+
+- 37c64d3: subintent support
+- e62bfdb: preview tx v2
+
+## 0.6.0-dev.1
+
+### Minor Changes
+
+- e62bfdb: preview tx v2
+
+## 0.6.0-dev.0
+
+### Minor Changes
+
+- 37c64d3: subintent support
+
 ## 0.5.0
 
 ### Minor Changes