+++ packages/tx-tool/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @radix-effects/tx-tool
 
+## 0.1.0
+
+### Minor Changes
+
+- 37c64d3: subintent support
+- e62bfdb: preview tx v2
+
+### Patch Changes
+
+- Updated dependencies [37c64d3]
+- Updated dependencies [e62bfdb]
+  - @radix-effects/gateway@0.6.0
+
+## 0.1.0-dev.1
+
+### Minor Changes
+
+- e62bfdb: preview tx v2
+
+### Patch Changes
+
+- Updated dependencies [e62bfdb]
+  - @radix-effects/gateway@0.6.0-dev.1
+
+## 0.1.0-dev.0
+
+### Minor Changes
+
+- 37c64d3: subintent support
+
+### Patch Changes
+
+- Updated dependencies [37c64d3]
+  - @radix-effects/gateway@0.6.0-dev.0
+
 ## 0.0.6
 
 ### Patch Changes